@@ -5,41 +5,33 @@
   "main": "index.js",
   "homepage": "./",
   "dependencies": {
-    "@ant-design/icons": "^4.5.0",
-    "@nivo/calendar": "^0.67.0",
-    "@nivo/core": "^0.67.0",
+    "@ant-design/icons": "^5.2.6",
+    "@nivo/calendar": "^0.83.0",
+    "@nivo/core": "^0.83.0",
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
     "@types/jest": "^26.0.15",
     "@types/node": "^12.0.0",
-    "@types/react": "^17.0.0",
-    "@types/react-collapse": "^5.0.0",
-    "@types/react-dom": "^17.0.0",
     "@types/react-helmet": "^6.1.0",
-    "@types/react-router": "^5.1.11",
-    "@types/react-router-dom": "^5.1.7",
-    "@types/react-toggle": "^4.0.2",
-    "antd": "^4.16.6",
-    "react": "^17.0.1",
-    "react-collapse": "^5.1.0",
-    "react-dom": "^17.0.1",
-    "react-github-button": "^0.1.11",
+    "antd": "^5.10.3",
+    "react": "^18.2.0",
+    "react-collapse": "^5.1.1",
+    "react-dom": "^18.2.0",
     "react-helmet": "^6.1.0",
-    "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.2",
-    "react-toggle": "^4.1.1",
-    "recharts": "^2.0.8",
-    "sass": "^1.32.8",
-    "typescript": "^4.1.2",
-    "web-vitals": "^1.0.1"
+    "react-router": "^6.17.0",
+    "react-router-dom": "^6.17.0",
+    "react-toggle": "^4.1.3",
+    "recharts": "^2.9.1",
+    "typescript": "^4.1.2"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "lint:prettier": "prettier --check 'src/**/*.{js,jsx,ts,tsx,css,scss,md}'",
+    "lint-fix:prettier": "prettier --write 'src/**/*.{js,jsx,ts,tsx,css,scss,md}'"
   },
   "eslintConfig": {
     "extends": [
@@ -59,10 +51,18 @@
       "last 1 safari version"
     ]
   },
-  "devDependencies": {
-    "prettier": "2.2.1"
-  },
   "resolutions": {
     "react-spring": "9.1.2"
+  },
+  "devDependencies": {
+    "@types/react": "^18.2.33",
+    "@types/react-collapse": "^5.0.3",
+    "@types/react-dom": "^18.2.14",
+    "@types/react-router": "^5.1.20",
+    "@types/react-router-dom": "^5.3.3",
+    "@types/react-toggle": "^4.0.4",
+    "prettier": "^3.0.3",
+    "react-scripts": "^5.0.1",
+    "sass": "^1.69.5"
   }
 }
